+++ src/main/java/com/spirnt/mission/discodeit/service/basic/BasicUserService.java
@@ -1,22 +1,14 @@
 package com.spirnt.mission.discodeit.service.basic;
 
-import com.spirnt.mission.discodeit.dto.binaryContent.BinaryContentCreate;
 import com.spirnt.mission.discodeit.dto.user.UserCreateRequest;
-import com.spirnt.mission.discodeit.dto.user.UserResponse;
 import com.spirnt.mission.discodeit.dto.user.UserUpdateRequest;
-import com.spirnt.mission.discodeit.dto.userStatus.UserStatusCreate;
 import com.spirnt.mission.discodeit.enity.User;
-import com.spirnt.mission.discodeit.enity.UserStatusType;
-import com.spirnt.mission.discodeit.repository.BinaryContentRepository;
-import com.spirnt.mission.discodeit.repository.ChannelRepository;
 import com.spirnt.mission.discodeit.repository.UserRepository;
-import com.spirnt.mission.discodeit.repository.UserStatusRepository;
-import com.spirnt.mission.discodeit.service.*;
+import com.spirnt.mission.discodeit.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.Instant;
 import java.util.*;
 
 @Service
@@ -24,15 +16,6 @@ import java.util.*;
 public class BasicUserService implements UserService {
     @Autowired
     private final UserRepository userRepository;
-//    @Autowired
-//    private ChannelService channelService;
-//    @Autowired
-//    private MessageService messageService;
-//    @Autowired
-//    private BinaryContentService binaryContentService;
-//    @Autowired
-//    private UserStatusService userStatusService;
-
 
     @Override
     public User create(UserCreateRequest dto) {
@@ -43,13 +26,7 @@ public class BasicUserService implements UserService {
             throw new IllegalArgumentException(dto.getName() + "Name Already Exists");
         }
         User user = new User(dto);
-//        //프로필 이미지 저장
-//        if (dto.getProfileImage() != null)
-//            binaryContentService.create(new BinaryContentCreate(user.getId(), null, dto.getProfileImage()));
         userRepository.save(user);
-//        // UserStatus 생성
-//        UserStatusCreate userStatusCreate = new UserStatusCreate(user.getId(), UserStatusType.ONLINE, Instant.now());
-//        userStatusService.create(userStatusCreate);
         return user;
     }
 
@@ -90,14 +67,11 @@ public class BasicUserService implements UserService {
     @Override
     public void delete(UUID userId) {
         // 존재하는 유저인지 검증
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NoSuchElementException("User ID: " + userId + " Not Found"));
-//        // 해당 유저가 속한 모든 채널에서 삭제하기 위해 채널 서비스 호출
-//        channelService.deleteUserInAllChannels(userId);
-//        // 유저의 프로필 삭제
-//        // UserStatus 삭제
-//        userStatusService.deleteByUserId(userId);
-        userRepository.delete(userId);
+        if (userRepository.existsById(userId)) {
+            userRepository.delete(userId);
+        } else {
+            throw new NoSuchElementException("User ID: " + userId + " Not Found");
+        }
     }
